Add Java 16 source sets and related tests for records
